:truck: Group test files
Moved all insert test files into a new package
called insert_tests.
