+++ src/tests/insert_tests/test_insert_engagement.py
@@ -1,20 +1,21 @@
 import unittest
-from sqlalchemy import select
 from aiounittest import async_test
-from . import get_test_database_session
-from . import create_all_tables
-from . import drop_all_tables
+from sqlalchemy import select
+from src.tests import get_test_database_session
+from src.tests import create_all_tables
+from src.tests import drop_all_tables
 from src.domain_logic.user_domain import UserDomain
 from src.domain_logic.blog_domain import BlogDomain
+from src.domain_logic.engagement_domain import EngagementDomain
 from src.usecases.insert.insert_user import insert_user
 from src.usecases.insert.insert_blog import insert_blog
-from src.model.blog import Blog
+from src.usecases.insert.insert_engagement import insertEngagement
+from src.model.engagement import Engagement
 
 
-class InsertBlog(unittest.TestCase):
-
+class InsertEngament(unittest.TestCase):
     @async_test
-    async def test_insert_blog(self):
+    async def test_insert_engagment(self):
         await create_all_tables()
         # insert user, id needed to create blog
         user_data = {"username": "Mako",
@@ -34,16 +35,17 @@ class InsertBlog(unittest.TestCase):
         blog_domain = BlogDomain(**blog_data)
         await insert_blog(blogDomain=blog_domain, func=get_test_database_session)
 
-        query = select(Blog).filter(Blog.id == 1)
+        engagement_data = {"blog_id": 1, "user_id": 1, "isLiked": True}
+        engagement_domain = EngagementDomain(**engagement_data)
+        await insertEngagement(engagementDomain=engagement_domain, func=get_test_database_session)
+
+        query = select(Engagement).filter(Engagement.id == 1)
         session = get_test_database_session()
+
         result = await session.execute(query)
-        blog: Blog = result.scalar_one_or_none()
+        engament: Engagement = result.scalar_one_or_none()
+        self.assertIsNotNone(engament)
+        self.assertTrue(engament.isLiked)
+        self.assertFalse(engament.isDisLiked)
 
-        self.assertIsNotNone(blog)
-        self.assertEqual(blog.title, "Async capabilities")
-        self.assertEqual(blog.body, "With async we can write apis that scale")
         await drop_all_tables()
-
-
-if __name__ == "__main__":
-    unittest.main()
